feat(home): retire dashboard page, feed becomes root
The feed view now lives at `/` and is labeled "Home" in the nav.
The old dashboard page (accounts, allocation bar, agent reports
card, recent transactions) is gone — its surfaces are covered by
/transactions, /reports, and /connections, and the feed is what
people actually open the app for.

- `/` serves FeedHandler with the onboarding redirect carried over
- `/feed` 301-redirects to `/` for any straggling links
- nav merges Home + Feed into one entry; the unread-reports badge
  rides the surviving Home item
- internal `/feed?...` hrefs (filter, before, clear) become `/`
- delete dashboard.go, dashboard.templ, dashboard_types.go,
  dashboard.js; move shared `relativeTime` helper into helpers.go
