+++ internal/templates/components/pages/errors.templ
@@ -14,7 +14,7 @@ templ NotFound() {
 		<p class="bb-error-message">The page you're looking for doesn't exist or has been moved.</p>
 		<a href="/" class="btn btn-primary rounded-xl px-6">
 			<i data-lucide="arrow-left" class="w-4 h-4"></i>
-			Back to Dashboard
+			Back to Home
 		</a>
 	</div>
 }
@@ -35,7 +35,7 @@ templ InternalError() {
 				Try Again
 			</a>
 			<a href="/" class="btn btn-ghost rounded-xl px-6">
-				Dashboard
+				Home
 			</a>
 		</div>
 	</div>
