+++ internal/templates/components/pages/feed.templ
@@ -40,7 +40,7 @@ templ feedActiveFilterBanner(p FeedProps) {
 	<div class="flex flex-wrap items-center gap-2 text-xs text-base-content/60 mb-3 px-1">
 		<i data-lucide="filter" class="w-3 h-3"></i>
 		<span>Showing only: <span class="font-medium text-base-content/80">{ feedFilterLabel(p.Filter) }</span></span>
-		<a href="/feed" class="text-primary/80 hover:text-primary no-underline ml-1">Clear filter</a>
+		<a href="/" class="text-primary/80 hover:text-primary no-underline ml-1">Clear filter</a>
 	</div>
 }
 
@@ -279,7 +279,7 @@ templ feedEmptyFiltered(filter string) {
 		<p class="text-sm text-base-content/55 mb-5">
 			Nothing in this slice today.
 		</p>
-		<a href="/feed" class="btn btn-sm btn-ghost rounded-xl border border-base-300">
+		<a href="/" class="btn btn-sm btn-ghost rounded-xl border border-base-300">
 			<i data-lucide="x" class="w-3.5 h-3.5"></i>
 			Clear filter
 		</a>
@@ -1250,13 +1250,13 @@ func feedHeroSubReports(p FeedProps) string {
 
 func feedFilterHRef(slug string) string {
 	if slug == "" {
-		return "/feed"
+		return "/"
 	}
-	return "/feed?filter=" + slug
+	return "/?filter=" + slug
 }
 
 // feedLoadOlderHRef builds the href for the "Load older activity" button:
-// `/feed?before=<oldestVisible>` plus the active filter (so an active chip
+// `/?before=<oldestVisible>` plus the active filter (so an active chip
 // survives pagination). The timestamp is RFC3339 so the handler's
 // `time.Parse(time.RFC3339, …)` round-trips it without any loss.
 func feedLoadOlderHRef(p FeedProps) string {
@@ -1264,7 +1264,7 @@ func feedLoadOlderHRef(p FeedProps) string {
 	if p.Filter != "" {
 		q = "filter=" + p.Filter + "&" + q
 	}
-	return "/feed?" + q
+	return "/?" + q
 }
 
 // feedFilterLabel renders the chip-name string used in the active-filter
